+++ CHANGELOG.md
@@ -11,9 +11,10 @@
   of the message body and do not count towards message length and other rules.
   Trailers are things like "Co-authored-by" and "Signed-off-by" lines at the
   end of the message body.
-- Add MessageCoAuthoredBy rule. This new rule checks if all "Co-authored-by"
-  lines are at the end of the message body. If they are not, GitHub and other
-  tools will not recognize the line and will not credit to the co-author.
+- Add MessageTrailerLine rule. This new rule checks if all "Co-authored-by",
+  "Signed-off-by" and "Helped-by" lines are at the end of the message body. If
+  they are not, GitHub and other tools will not recognize the line and will not
+  credit to the co-author or recognize the sign off.
 
 ## Changed